Refactor plugin code base
- The rateLimit class stores timeLapse as an integer.
  PeriodicRateLimiter stores timeLapse as Long, which confuses
  understanding and using the code. This change converts all timeLapse
  variables to an integer.
  
- Remove timeLapse from warning rate limiters, since timeLapse can be
  accesed from delegate
  
- Add methods getWarnLimit and getTimelapse

+++ src/test/java/com/googlesource/gerrit/plugins/ratelimiter/PeriodicRateLimiterTest.java
@@ -71,8 +71,8 @@ public class PeriodicRateLimiterTest {
     verify(scheduledExecutorMock)
         .scheduleAtFixedRate(
             any(),
-            eq(DEFAULT_TIME_LAPSE_IN_MINUTES),
-            eq(DEFAULT_TIME_LAPSE_IN_MINUTES),
+            eq((long) DEFAULT_TIME_LAPSE_IN_MINUTES),
+            eq((long) DEFAULT_TIME_LAPSE_IN_MINUTES),
             eq(TimeUnit.MINUTES));
   }
 
@@ -82,8 +82,8 @@ public class PeriodicRateLimiterTest {
     verify(scheduledExecutorMock)
         .scheduleAtFixedRate(
             runnableCaptor.capture(),
-            eq(DEFAULT_TIME_LAPSE_IN_MINUTES),
-            eq(DEFAULT_TIME_LAPSE_IN_MINUTES),
+            eq((long) DEFAULT_TIME_LAPSE_IN_MINUTES),
+            eq((long) DEFAULT_TIME_LAPSE_IN_MINUTES),
             eq(TimeUnit.MINUTES));
 
     // Use all permits
